requirements.txt
python-telegram-bot==12.6.1
pandas==1.0.3
numpy==1.18.1
matplotlib==3.1.3
requests==2.22.0